backend/pythonserver/requirements.txt
certifi==2020.12.5
chardet==4.0.0
dnspython==2.1.0
facebook-sdk==3.1.0
idna==2.10
oauthlib==3.1.0
pymongo==3.11.3
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
tweepy==3.10.0
urllib3==1.26.3
